Feature/planner trouble shooting

+++ app/src/main/java/ajou/paran/entrip/repository/network/PlannerRemoteSource.kt
@@ -1,42 +1,35 @@
 package ajou.paran.entrip.repository.network
 
+import ajou.paran.entrip.model.PlanEntity
 import ajou.paran.entrip.model.PlannerEntity
 import ajou.paran.entrip.repository.network.api.PlanApi
-import ajou.paran.entrip.repository.network.dto.planner.CreatePlannerDto
-import ajou.paran.entrip.repository.network.dto.planner.UpdatePlannerRequestDto
+import ajou.paran.entrip.repository.network.dto.PlannerUpdateRequest
 import ajou.paran.entrip.util.network.BaseResult
 import ajou.paran.entrip.util.network.Failure
 import ajou.paran.entrip.util.network.networkinterceptor.NoInternetException
+import android.util.Log
 
-class PlannerRemoteSource
-constructor(private val planApi: PlanApi)
-{
+class PlannerRemoteSource constructor(private val planApi: PlanApi) {
     /**
      * @POST : save(2ntrip.com/api/v1/planners)
      * @RequestBody : String user_id
      * **/
-    suspend fun createPlanner(user_id: String): BaseResult<PlannerEntity, Failure>
-            = try {
-        val obj = CreatePlannerDto(user_id)
-        val response = planApi.createPlanner(obj)
-        if (response.httpStatus == 200) {
-            val planner = response.data.let { t ->
-                PlannerEntity(
-                    t.planner_id,
-                    t.title,
-                    t.start_date,
-                    t.end_date,
-                    t.timeStamp
-                )
+    suspend fun createPlanner(userId : String): BaseResult<PlannerEntity, Failure> {
+        try {
+            val response = planApi.createPlanner(userId)
+            return if (response.status == 200) {
+                val planner = response.data?.let { t ->
+                    PlannerEntity(t.planner_id, t.title, t.start_date, t.end_date, t.timeStamp, t.comment_timeStamp)
+                }
+                BaseResult.Success(planner!!)
+            } else {
+                BaseResult.Error(Failure(response.status, response.message))
             }
-            BaseResult.Success(planner)
-        } else {
-            BaseResult.Error(Failure(500, response.message))
+        } catch (e: NoInternetException) {
+            return BaseResult.Error(Failure(0, e.message))
+        } catch (e: Exception) {
+            return BaseResult.Error(Failure(-1, e.message.toString()))
         }
-    } catch (e: NoInternetException) {
-        BaseResult.Error(Failure(0, e.message))
-    } catch (e: Exception) {
-        BaseResult.Error(Failure(-1, e.message.toString()))
     }
 
     /**
@@ -44,28 +37,23 @@ constructor(private val planApi: PlanApi)
      * @PathVariable : Long planner_id
      * @RequestBody : String title, String start_date, String end_date
      * **/
-    suspend fun updatePlanner(plannerId: Long, planner: PlannerEntity): BaseResult<PlannerEntity, Failure>
+    suspend fun updatePlanner(plannerId: Long, planner: PlannerUpdateRequest): BaseResult<PlannerEntity, Failure>
             = try {
-        val response = planApi.updatePlanner(
-            plannerId,
-            UpdatePlannerRequestDto(
-                title = planner.title,
-                start_date = planner.start_date,
-                end_date = planner.end_date
-            )
-        )
-        if (response.httpStatus == 200) {
+        val response = planApi.updatePlanner(plannerId,planner)
+        if (response.status == 200) {
             val updatePlanner = response.data.let { t ->
                 PlannerEntity(
                     planner_id = t.planner_id,
                     title = t.title,
                     start_date = t.start_date,
                     end_date = t.end_date,
-                    timeStamp = t.timeStamp)
+                    time_stamp = t.time_stamp,
+                    comment_timeStamp = t.comment_timeStamp
+                )
             }
             BaseResult.Success(updatePlanner)
         }else{
-            BaseResult.Error(Failure(500, response.message))
+            BaseResult.Error(Failure(response.status, response.message))
         }
     }catch(e: NoInternetException){
         BaseResult.Error(Failure(0, e.message))
@@ -77,40 +65,39 @@ constructor(private val planApi: PlanApi)
      * @GET : findById (2ntrip.com/api/v1/planners/{planner_id})
      * @PathVariable : Long planner_id
      * **/
-    suspend fun findPlanner(plannerId: Long): BaseResult<PlannerEntity, Failure>
-            = try {
-        val response = planApi.findPlanner(plannerId)
-        if (response.httpStatus == 200) {
-            val planner = response.data.let { t ->
-                PlannerEntity(
-                    planner_id = t.planner_id,
-                    title = t.title,
-                    start_date = t.start_date,
-                    end_date = t.end_date,
-                    timeStamp = t.timeStamp
-                )
+    suspend fun findPlanner(plannerId: Long): BaseResult<PlannerEntity, Failure> {
+        try {
+            val response = planApi.fetchPlanner(plannerId)
+            return if (response.status == 200) {
+                val planner = response.data?.let { t ->
+                    PlannerEntity(t.planner_id, t.title, t.start_date, t.end_date, t.timeStamp, t.comment_timeStamp)
+                }
+                BaseResult.Success(planner!!)
+            } else {
+                BaseResult.Error(Failure(response.status, response.message))
             }
-            BaseResult.Success(planner)
-        } else {
-            BaseResult.Error(Failure(500, response.message))
+        } catch (e: NoInternetException) {
+            return BaseResult.Error(Failure(0, e.message))
+        } catch (e: Exception) {
+            return BaseResult.Error(Failure(-1, e.message.toString()))
         }
-    } catch (e: NoInternetException) {
-        BaseResult.Error(Failure(0, e.message))
-    } catch (e: Exception) {
-        BaseResult.Error(Failure(-1, e.message.toString()))
     }
 
+
     /**
-     * @GET : plannerIsExistWithId(2ntrip.com/api/v1/planners/{planner_id}/exist)
-     * @PathVariable : Long planner_id
-     * **/
+     *  Home 화면 -> Planner 로 넘어가는 과정에서 사용자가
+     *  planner list 중 planner를 선택했을 때 호출되는 함수
+     *  isExist -> True : 해당 planner가 서버에 있으므로 planView로 넘어가서 sync작업
+     *          -> False : 해당 planner가 서버에 없으므로 planView로 넘어가지 않고,
+     *                     Home 화면에 Planner가 없다고 알려줘야 한다.
+     */
     suspend fun isExist(planner_id: Long): BaseResult<Boolean, Failure>
             = try {
         val response = planApi.isExist(planner_id)
-        if (response.httpStatus == 200) {
+        if (response.status == 200) {
             BaseResult.Success(response.data)
         } else {
-            BaseResult.Error(Failure(500, response.message))
+            BaseResult.Error(Failure(response.status, response.message))
         }
     } catch (e: NoInternetException) {
         BaseResult.Error(Failure(0, e.message))
@@ -125,15 +112,64 @@ constructor(private val planApi: PlanApi)
      * **/
     suspend fun deletePlanner(planner_id: Long) : BaseResult<Long, Failure>
             = try{
-        val response = planApi.deletePlanner(planner_id)
-        if (response.httpStatus == 200){
-            BaseResult.Success(response.data)
-        } else {
-            BaseResult.Error(Failure(500, response.message))
+        val response1 = planApi.isExist(planner_id)
+        var isExist : Boolean
+        if(response1.status == 200){
+            isExist = response1.data
+            if(isExist){
+                val response2 = planApi.deletePlanner(planner_id)
+                if(response2.status == 200) BaseResult.Success(response2.data)
+                else BaseResult.Error(Failure(response2.status, response2.message))
+            }else{
+                // 존재하지 않는 경우(다른 사용자가 이미 삭제함)
+                BaseResult.Success(planner_id)
+            }
+        }else{
+            BaseResult.Error(Failure(response1.status, response1.message))
         }
     } catch (e: NoInternetException) {
         BaseResult.Error(Failure(0, e.message))
     } catch (e: Exception) {
         BaseResult.Error(Failure(-1, e.message.toString()))
     }
+
+    suspend fun fetchPlans(planner_idFK: Long): BaseResult<List<PlanEntity>, Failure> {
+        try {
+            val response = planApi.fetchPlans(planner_idFK)
+            return if (response.status == 200) {
+                val plans = mutableListOf<PlanEntity>()
+                response.data?.forEach { t ->
+                    plans.add(
+                        PlanEntity(t.id, t.planner_idFK, t.todo, t.rgb, t.time, t.location, t.date)
+                    )
+                }
+                BaseResult.Success(plans)
+            } else {
+                BaseResult.Error(Failure(response.status, response.message))
+            }
+        } catch (e: NoInternetException) {
+            return BaseResult.Error(Failure(0, e.message))
+        } catch (e: Exception) {
+            return BaseResult.Error(Failure(-1, e.message.toString()))
+        }
+    }
+
+    suspend fun fetchPlanner(planner_id: Long): BaseResult<PlannerEntity, Failure> {
+        try {
+            val response = planApi.fetchPlanner(planner_id)
+            return if (response.status == 200) {
+                val planner = response.data?.let { t ->
+                    PlannerEntity(t.planner_id, t.title, t.start_date, t.end_date, t.timeStamp, t.comment_timeStamp)
+                }
+                BaseResult.Success(planner!!)
+            } else {
+                BaseResult.Error(Failure(response.status, response.message))
+            }
+        } catch (e: NoInternetException) {
+            return BaseResult.Error(Failure(0, e.message))
+        } catch (e: Exception) {
+            return BaseResult.Error(Failure(-1, e.message.toString()))
+        }
+    }
+
 }
